backend/requirements.txt
adjustText==0.7.3
asgiref==3.5.2
beautifulsoup4==4.11.1
blog==1.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
decorator==5.1.1
Django==3.2.13
djangorestframework==3.13.1
docopt==0.6.2
drf-yasg==1.20.0
environs==9.5.0
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
item==0.0.3
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
limit==0.2.3
many==0.6.9
Markdown==3.3.7
MarkupSafe==2.1.1
marshmallow==3.16.0
matplotlib==3.5.2
numpy==1.21.6
objects==0.3.1
packaging==21.3
pandas==1.1.5
patsy==0.5.2
Pillow==9.1.1
psycopg2==2.9.3
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.12.0
pymdown-extensions==9.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytils==0.4.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selection==0.0.14
self==2020.12.3
SHop==0.0.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
statsmodels==0.13.2
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
user-agent==0.1.10
weblib==0.1.30
zipp==3.8.0
